Unify Qwen2RMSNorm definitions and use RMSNorm from PyTorch
